+++ src/main/java/com/springapp/mvc/UserRepository.java
@@ -2,5 +2,7 @@ package com.springapp.mvc;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+
 }
